Easy-but-not-correct reliability improvement
Sometimes this test fails because the timing is not correct. We blindly
sleep for a second to check the snapshot file because that is the
setting that the server was started with. Unless working on a real-time
system one cannot really rely much on timings alone to avoid race
conditions though. This commit simply makes the sleep greater to have a
lower probability of failing, but doesn't mean we are immune to failure.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>