Installing berkeley-db@4
The more generic berkeley homebrew formula now migrated to use
berkeley-db 18.1, which is not supported yet by the bsddb3 python
package (which, on the other hand, we need to install in MacOS). We are
thus using berkeley-db@4, which installs 4.8.X, which IS supported by
the latest bsddb3.

This should fix the travis failure we are currently experiencing on
our MacOS platform.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>