Refactoring sending of files
Until now, we have always sent one file per response. This is about to
change with the introduction of tarfile responses, where multiple files
need to be sent through the same response, meaning that sendfile will be
used more than once. It also means that the sending of the initial
headers advertising the single file to the client can be refactored into
its own method to make it more reusable.

The net result of this is that the old send_file method's body has been
refactored into the new send_file_headers and write_file_data methods,
which take care of each corresponding bit of functionality. The latter
is called write_file_data instead of send_file_data to maintain
consistency with the rest of the send_* methods, all of which initiate
an HTTP response (while write_file_data doesn't). Similarly, a new
write_data method has been spawned from the old send_data method.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>