NGAS-139 retrieve container contents as a tarfile
This implementation performs no compression on the data, and therefore
we can benefit from using sendfile. This means we have a mix of writes,
with the tar headers being written from memory, and the files' data
being sendfile'd. The tar headers are calculated using the TarInfo class
from the tarfile system module. The writing is not done through the
TarFile class though, because it internally buffers writes without the
possibility of flushing.

On the client side we use extractall() to extract the tarfile contents
into a particular directory.

Of course, new unit tests double-check that this is working as expected.
I also used the opportunity to simplify a bit the existing tests and
port them to use the new assert_ngas_status method.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>