Built-in commands to commands subpackage
Not only are we moving the different modules for the built-in NGAS
commands, but we are also renaming them to a simpler naming scheme
without the duplicated "ngams" and "Cmd" prefixes. Their new names are
exactly the lowercase version of the command they serve, and we use this
property to simplify the command module loading logic.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>