Moving janitor modules down to ngamsServer
When we broke down the janitor thread logic into separate modules we
put the new modules under the ngamsPlugIns package. In hindsight this
looks like a pretty bad decision, since this code is a core part of the
ngamsServer module. With this commit we are bringing them back, and
taking also the opportunity the move the rest of the janitor-related
modules into the janitor subpackage.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>