Bye bye, TRACE()
The TRACE() object was a part of the old home-brewed logging system,
and allowed to print the beginning and the end of the execution of the
function where it was declared, together with the function duration.
When we removed the old logging system and replaced it with the standard
python logging module (back in version 9.0), I did not have the heart to
remove *all* the calls to TRACE immediately. Two major versions down the
line now, I think it's time to remove them.

If in the future we need a way to print in the logs some function
tracking information, it could probably be done with a decorator.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>