Increasing authentication/authorization coverage
First of all, coverage information for the authentication and
authorization logic in NGAS was almost fully missing. This was due to
the server being shut down via SIGKILL by the test classes, which was
caused by unauthenticated STATUS/OFFLINE/EXIT requests from the client
side. We previously added support for authenticated server shutdowns, so
we are using that now to cleanly shut the servers down.

With proper coverage information we also found out that some of the
lines we *thought* we were testing were actually not being hit. This was
due to a wrong logic in the _assert_code method which wrongly grouped
conditions and statements. It has been fixed now, and coverage is now as
it was intended to be.

Additionally, we added a new unit test for things we were not previously
checking before, like unknown user and password usage. This is bringing
up the coverage of the ngamsAuthUtils module up to 100%.

Finally, some general cleanup and code reorganization happened in the
test class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>