Reorganizing test support modules
Several of the modules present in the ngamsTest package are not tests
themselves, but support modules that are needed by some of the unit
tests. In order to have a clearer structure I'm moving them to a new
support subpackage, from where they can still be safely loaded.

During this process I also identified several of these modules that were
not being used anymore, so I've simply removed them. As a result of this
I'll be able to remove some of the test_* methods present on the
ngamsServer class which serve as an anchor point for test server classes
to kill the server at particular points in time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>