Removing test_* methods from ngamsServer
These were used by test classes that inherited from ngamsServer to kill
the server at specific times. These classes are not used currently, and
therefore the existence of these methods is not necessary at all.

There is still one test_* method needed by the unit tests, but in the
future we should find out how to re-implement this without having to
have test_* methods in the server class itself.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>