Removing unuseful ngamsTestAsyncListRetrieve tests
These tests were never quite well integrated with the rest of the unit
tests, and were also testing some functionality that is not quite well
integrated with NGAS itself. I think it's thus simpler to remove them
for the time being, and re-insert them properly as unit tests if we
continue properly supporting the ASYNCRETRIEVE command.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>