Tests moved to test/ top-level directory
This is it. For a long time the NGAS unit tests were buried inside the
deep bowels of the src/ngamsTest/ngamsTest directory. This changes
today, with them now being found directly under the test/ directory at
the top level of the repository. This makes it easier than ever to run
the tests from the repository root directory using standard unit test
running tools.

We still need to move ourselves *into* the test/ directory to run the
tests, because all relative paths to static files used by the tests are
based on this. One day we might be able to perfect this, maybe using
pkg_resources or __file__, but for the time being we use this shortcut.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>