Adding a Changelog
The contents of this changelog are based on the contents of the "NGAS
Blog" on the NUG confluence site, and the per-version tickets on the
NGAS JIRA project. Both of these are mostly based on the git history of
the project (specially for the list of features in old releases), which
is the ground truth.

This is only documentation, so let's [skip ci].

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>