New Task class to handle background tasks
The NGAS server handles a number of different background tasks, all of
which need to be stopped when the server comes down. Until the code
managing these tasks has not been unified, leading to maintenance
difficulties and code duplication.

This commit introduces the new code that should be common to all these
background tasks. It includes support for thread-or-process background
tasks, and it takes care of creating and passing down the necessary
stopping events. In the future this could be further refined to pass
down suspend and check functions instead of the stop event. This way the
target functions don't need to pass down the event and have their own
checking/suspend functions (which are, again, duplicated throughout
different modules), but simply call the necessary function without
worrying how it is implemented underneath.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>