Porting background tasks to use Task class
A few small things had to be adjusted, like accommodating with the fact that
the stop event is the last *args passed to the Task's target function.

This removes a lot of unnecessary and repeated code from the ngamsServer
class, making it far more maintainable. It also reduces the usability of
the thread_run_permission flag (i.e., get/setThreadRunPermission
methods) on the ngamsServer. This flag was used widely not only by some
of the background tasks, but also by the asynchronous version of the
CLONE command. Since the flag is now used *only* by the asynchronous
version of CLONE I simplified it to be a simple attribute, and renamed
it to run_async_commands.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>