Using stop event in mirroring control thread
Previously this code was using plain sleeps in places were doing a wait
on the stopping event should have happened instead. I'm correcting this,
and therefore bringing down the mirroring thread now happens almost
immediately.

The mirroring sub-threads also need to be taken into account. They can
be waiting on a "trigger" event that lives on the server, and can hung
up there even if the main mirroring thread finishes. To solve this at
the moment I'm setting that event, so the sub-threads continue running,
and exit after realizing the stop event has been set too.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>