Fixing few bugs in QUERY command
Because map()s in python 3 can only be iterated once, there was a
problem with the generation of the text that was sent to the client
after performing more than one query with the list and json formats,
since they both iterated over the column names to include them in the
resulting text.

A unit test is being added now to make sure this doesn't happen again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>