Support for reading file compression column
In order to properly implement a fix for NGAS-73 we need to get a hold
on the compression format under which the data is stored in NGAS. This
information will then be used by the RETRIEVE command to properly
handle the response back to the client.

The RETRIEVE command, for historical reasons, supports two ways of
querying the information about a particular file, and therefore this
commit adds support for both these mechanisms. The two calls used by
RETRIEVE are used in some other places in the code, so care has been
taken to keep the interface backwards-compatible by adding a new
optional argument to signal the inclusion of the compression information
on the returned structures, with a default value of False.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>