send_file/send_file_headers support extra headers
The send_file_headers creates its own set of HTTP headers than then are
sent to the client. On top of those we'll need to send some extra
headers in the future, so we are adding support for that. Still, the
headers created internally by send_file_headers take precedence over the
ones given by external callers in case of a conflict.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>