NGAS-73 properly sending back compressed content
NGAS-73 describes how archived NGAS logs are not correctly returned by
the server, in the sense that their content is gzip'ed, but the filename
that ends up being written in the disk (i.e., the filename written in
the Content-Disposition header) doesn't contain a '.gz' extension. This
obviously causes confusion to tools like graphical file explorers, which
associate file extensions with programs to handle specific types of
files. To properly open these files users had to rename the returned
file to something containing the '.gz' extension, which enabled the
proper program to be opened to deal with the compressed content.

The fix implemented on this commit fixes this problem in two different
ways. First, if the client supports gzip'ed content encoding for the
HTTP response (i.e., they send the Accept-Encoding header with 'gzip' in
its value) then we leave the filename untouched, but set the
'Content-Encoding' header in the response to 'gzip'. This will cause the
compressed content to be automatically decompressed by the client before
writing to disk, so the file on disk will be decompressed. On the other
hand, if the user doesn't handle gzip content automatically, then we
make sure the filename communicated in the Content-Disposition header
contains a final '.gz' extension.

Note that in both cases we still send the content compressed; the
different lays on whether the user will have to decompress the content
manually or automatically.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>