Using better assertions on tests
This change is two-fold: on the one hand, we are getting rid of many
warning messages due to us using the assertEquals method of the
unittest.TestCase class instead of assertEqual (the former is deprecated
and will eventually be removed, while the second has always been
present). On the other hand, it was a good opportunity to go through the
code and start using our assert_ngas_status, assertArchive and
assertStatus methods, which make it easier for users to make sure
requests sent to the server actually ran fine without too much hassle.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>