Correct and more compact .travis.yml
First, python 3.7 is not available in trusty images (but is available in
xenial). On the other hand we had a lot of duplication in our matrix
declaration, so we are hoping that this commit will make it more maintainable.
Sadly, the global and local apt addons package lists don't get merged
automatically, so we need to repeat the bsddb development package when
necessary; likewise for the hostname addon.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>