Temporary print to debug problem in Travis MacOS
For some reason connected to us setting a hostname via the hostname
travis addon in MacOS jobs (we previously didn't) all the idle
suspension tests are now failing due to invalid keys in the cluster
dictionary being returned by prepSimCluster. This print should give us
an idea of what is going on, and how to fix it, which hopefully will not
involve *not* having to set the hostname in MacOS builds.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>