Removing unnecessary XML content replacement
The comment in the code read about the incapability of the Expat parser
to handle <?xml:stylesheet?> elements at an XML file's beginning. It
turns out that xml:stylesheet doesn't exist, but xml-stylesheet does,
which (I guess) was what caused the confusion in the first place and led
to the faulty string substitution to be put there.

I came around this bit of code due to the warnings that python 3.7 is
giving us about invalid escape sequences in regular expressions
patterns. With the code gone the warning is also gone. I also added a
small unit test that checks that xml-stylesheet (i.e., the correct form)
is correctly parsed, just in case.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>