Fixing regular expression pattern in db query
The regular expression to change '*' characters into '%' characters
before building the SQL query in the getFileInfoList method was wrong,
as it didn't properly handled the two-backslash problem of regular
expressions with simple strings. Using a raw string is simpler and gets
the work done.

To make sure this is working as intended a new unit test exercises the
method. There was no db-specific testing module, so I created a new one
where we can hopefully place db-specific tests in the future.

The warning was spotted by python 3.7's re module.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>