Invalid regular expression pattern fix
The regular expression used to search for invalid filename characters on
strings was faulty because it didn't properly escape the '?' character
actually (due to the usual string escaping v/s re escaping gotcha).
Python is clever enough to go around this, but in 3.7 we start getting
warnings, so let's avoid them.

In this commit I'm not only fixing the escaping of the '?' character,
but I'm also changing the re pattern (to use [] instead of many |), and
moving the replacing operation into its own method for easier reuse
(which we should, in some places) and testability.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>