Adding necessary mock modules to build API docs
The API docs built in our local development environment because we had
all the dependencies installed already, but it was actually failing to
build on RTDs because the different modules could not be imported. By
properly mocking the dependencies we can successfully import the
autodoc-ed modules and produce their documentation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>