Default to not doing on-stream checksuming
This is actually the safest path. With the previous default setting we
trusted the checksum calculated on the incoming stream as the final
value, even when later on we used a plug-in that did not state whether
it changed the data (and therefore, its checksum) or not. The safer
approach is to be conservative and be ready for the worst-case scenario.
Users who want to take advantage of the on-stream checksum calculation
therefore need to use a plug-in that explicitly promises NGAS not to
change the data.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>