Simpler error handling during archive processing
The previous code was unnecessarily complex, in part only because of the
log statement that would end up being added to the log stream. This was
unnecessary since the top-level request handling code already prints a
nice traceback whenever there is an exception being thrown from down
below. There was also a bit of duplication on how files from the staging
area were being cleaned up, which has been simplified and unified as
well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>