Highly improving our support for overwriting files
The previous version of the archiving code that dealt with overwriting
files was greatly broken, and also distributed across many places in the
code. Something needed to be done about it.

Some of the previous commits dealt with the latter, bringing the
different logic bits into the ngamsArchiveUtils module.  This
centralizes the maintenance of this logic, making it easier to follow
and maintain it. In a second round of previous commits we also set the
ground required to properly implement some of the features of file
overwriting. This included mostly proper database updates for
files/disks.

This commit brings all the pieces together. When we find an overwriting
request we now properly check if the request is valid (in terms of its
parameters), correctly deal with the existing database records for the
existing file and the original and new hosting disks (which might be
different), and also remove the previous copy of the data when
necessary. These expectations align with what we have in our new
(currently disabled by default) unit tests, which check that
archiving/retrieval requests pass/fail as corresponding, that files
contents are correctly overwritten, and that the database is correctly
updated every time.

Please note that the support for overwriting files has been restricted
to files that are archived locally. Dealing with remotely stored files
is even more complicated, so we are denying such requests to begin with.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>