Improving documentation about versioning
The ARCHIVE/QARCHIVE commands documentation did not specify how they
handled versioning. Since we have been working on this lately it makes
sense to use our fresh memories and dump them into a document.

I'm also taking the opportunity to add some more missing documentation
on some of the parameters used by these commands, and to update the text
and rearrange the contents so they better reflect the current state of
the code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>