Reading file_version only as needed
When overwriting we can specify a specific file version to use, but when
using versioning this makes no sense, and therefore we shouldn't be even
trying to read the file_version parameter. This, together with the
recent enhancements to the ngamsReqProps class, make the code much
easier to follow.

There was a unit test that actually relied on file_version being parsed
even for versioned archive operations (though only to double-check that
it failed to parse non-integer values), so I simply removed it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>