Don't use checksum plug-ins anymore
There is absolutely no need to use the old concept of a checksum plug-in
here. We use instead the newer CRC variant mechanism, which already
correctly and efficiently implements checksum calculations.

Some unit tests were relying on the old plug-in name appearing in the
responses, so those have been updated.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>