Transactional disk updates for file overwriting
When a file overwriting happens we need to transactionally update one or
two rows, depending on whether the new copy of the file sits on the same
disk where the old copy was or not. This new set of queries allows that.
They are designed to be called either independently, or as part of a
single transaction.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>