UPDATEs on ngas_files can move files across disks
This will allow us to properly update the DB record of file that are
being overwritten by a user request, but which contents sit on a
different (but still local) disk. We are allowing this through a new
optional argument though, so the default behavior of maintaining a file
within its disk remains.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>