A few improvements on crc_benchmark
The previous version was already much better in terms of features
compared to our old code, but it had two small problems. One was that
files as data sources for checksumming were not working anymore, so
we've fixed that. The other was that multiprocessing.RawArray.raw and
both io.BytesIO/StringIO trigger a copy of the data, and therefore some
of the setup times were unnecessarily too high. We now avoid all data
copying, which is specially important when using processes with shared
memory.

This is not part of the core NGAS code (should probably be moved
somewhere else even...) so we can safely [skip ci].

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>