Moving benchmarking scripts out of NGAS
These scripts are indeed useful for assessing NGAS installations, but
they are way more general than that and can be useful in other contexts,
so having them available as a separate project/repository in an
easy-to-find fashion is definitely a good thing. On top of that they
don't actually follow the NGAS release cycle, so it does not make sense
to version them together.

Based on that, I'm moving these scripts out of NGAS, and into a new
https://github.com/ICRAR/benchmark_tools repository.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>