Asserting without a reference file
These tests' reference files are quite simplistic, so the assertion they
are supporting doesn't need them in the first place. It's much easier in
these cases to simply use self.assert_ngas_status and be done with it.
There's surely more of these out there, but this is a good start.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>