Don't create directories in the ngamsConfig class
The previous code was way too eager, as almost every time a
configuration file was being loaded the code of the ngamsConfig class
would check for the existence of a few directories and create them in
case they didn't exist. Many times we just load configuration files
simply to inspect values (like, for example, during the unit tests), so
performing this check and creating these directories in the filesystem
during configuration loading was way too eager.

The directories included are the directory where the log file is hosted,
the processing directory and the root directory. These are created by
the ngamsServer class now instead, which is the one that actually needs
them there.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>