Proper, central processing directory calculation
For some reason the getProcessingDirectory method of the ngamsConfig
class was only calculating the "base" directory on which a "processing"
directory (the *actual* processing directory) had to be looked for.
Clients of this method therefore needed to add the final "processing"
path element to get what they were looking for.

This calculation has now been centralized, and the sample configuration
file has received a note about this.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>