New tmp_root global variable and utility functions
The tmp_root global variable will point to the directory that will be
the base of *all* temporary files created by unit testing code. These
files are currently split in two groups: things that go under ./tmp
(relative the the test/ directory), and things that go under
/tmp/ngamsTest, including all the NGAS root directories, and their
subdirectories, used by the servers spawned by the unit tests.  The
tmp_root global variable will default to the temporary folder on a
per-system basis (as returned by tempfile.gettempdir()), but users can
specify their own as well.

Additionally we created a handful of methods that will allow unit test
classes to easily refer to files inside the tmp_root directory, inside
an NGAS root directory (without having to pass down all the information
around), and to create NGAS disk IDs (which are basically the absname of
the volume on the filesystem with slashes replaced by dashes).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>