Unit tests using flexible temporary directory
There is no easy way to do this one: the change affects almost all unit
testing modules (if not all) and the reference files as well, so this
commit is quite big. In general I've changed all references to files
under "/tmp" or "./tmp" to use the flexible root directory that we now
support, and switched to using variables like %NGAS_ROOT% in the
reference files that are then replaced at runtime after loading the
reference file, but before performing the comparison.

Since we were jumping into a change-everything adventure, I also took
the change to simplify a bit how reference files were being asserted
against. Most times users needed to process the resulting data (out of
status objects, messages back from the server, etc) through a number of
filtering processes, and then compare with with what was stored in a
reference file. This way of chaining calls was duplicated everywhere,
and therefor it made the maintenance of these cases more complex than
necessary. The code *within* those filtering methods was also somewhat
duplicated, so I: a) centralised the filtering code, and b) went through
all unit tests and simplified how this filtering and matching was being
done. I also got rid of other unnecessary methods whose usage has been
removed.

In general there is much less code for the same amount of checking, and
also less interaction with the filesystem, so all in all it's a big win
for everyone.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>