Cache the hostname
There is no point on constantly querying the hostname from the system;
instead query it once and reuse the same value throughout the entire
execution. This obviously means that the code is not able to react to a
change in the hostname, but if that ever happens we'll have way bigger
problems to worry about.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>