LARCHIVE rewritten and made a first-class citizen
The LARCHIVE command previously consisted on a lot of code copied over
from ngamsArchiveUtils. This code duplication made it difficult to
maintain, and therefore it was never really looked after. This
refactoring (based on the previous two or three commits) greatly reduce
the amount of code in the module implementing the command, and moves it
to the ngamsServer.commands package with the rest of the built-in
commands.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>