self.get_client() in favour of sendPclCmd()
The latter is an old name that has stayed for a while and doesn't really
represent what is does: creating an ngas client and returning it to the
user. Having this new method as part of the base unit test class also
means we don't need to import an additional name from the ngamsTestLib
module.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>