Adding back 'processing' path element
This additional path element, added in c64803a2, was incorrectly removed
in 92061532. This removal produces the removal of the entire NGAS root
directory if the processing directory, as configured by the user, is '.'
(which is the value we show in the sample configuration file). This
obviously produces massive loss of data, and in fact this bit me in my
local testing NGAS root area.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>