Simplifying configuration checks
The way the code in the ngamsConfig was performing the checks of each
individual element of the configuration was way too verbose. Too many
calls were being done to get the reference of *exactly* the same
underlying object (a list of errors), which was then *always* passed as
a parameter to a few functions that were prepared for cases when the
parameter was not passed. These two points incremented the complexity of
the code unnecessarily.

In this commit I moved these functions within the ngamsConfig class so
they didn't need to get the list of errors as a parameter anymore
(because it can be found via self). This and a few other simplifications
make the code much nicer to look at, and shove away some ~120 lines of
unnecessary code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>