Creating default clients in ngamsTestSuite
Whenever a unit test starts up a server, it also usually needs a client
to that server. We have been creating these clients individually on each
unit test method, which is way too repetitive. Instead, we are now
always creating a client for the servers we start.

The clients are retrieved via self.client. In the common scenario of a
single server being started the self.client attribute will be the client
to that server, whatever port it is running on. If more than one server
is started then self.client is a callable that takes a single port
argument, and yields the client for that port.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>