Correct test expectations
This test has *always* (at least since the creation of the repository in
git -- ~2012) been incorrectly checking this particular "tag": instead
of passing the tag as a list with a single string, the tag was being
passed as a single string only, and therefore what was being checked was
that all characters in the string could be found somewhere in the log.
This was obviously not too hard to accomplish, and therefore the test
was giving a false sense of security.

After correcting this, the test failed because the contents of the
reference file could indeed not be found in the logfile. I've adjusted
the reference file for the time being.

On that last poing, I have the impression that it should be the other
way around: the CLUSTER denomination of the host/file is probably what
we should get, but the code in ngamsServer says otherwise: the cluster
field of the host info object is filled with the "host id" and not with
the hostname. This has been the case even from before I moved getHostId
from its global context to be within the server object. The original
global getHostId returned either "<hostname>" or "<hostname>:<port>"
depending on whether one or more server were started, so two servers
started on a cluster probably effectively yielded two different cluster
names even back then. In any case the whole LOCAL/CLUSTER/DOMAIN/REMOTE
logic should probably be revisited, as it's a bit confusing at best, and
probably quite buggy in general.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>