More generic client asserting functions in tests
Based on self.archive, this new commit gives a more generic way to unit
test classes to easily invoke client methods on the pre-built clients and
assert the returned status automatically. This removes most of the
manual invocations to self.assert_ngas_status, which is now only needed
for a few instances where a more specialized client needs to be used.

The new methods come in two flavours: asserting for success or failure.
When a method is invoked with its normal name (e.g., "self.archive")
then a successful status is expected. In contrast, if it's invoked with
a "_fail" suffix (e.g., "self.retrieve_fail") then a failure is
expected. This removes the need to add extra parameters to the function
call from the unit test functions, which keeps the unit test code easier
to read. These two flavours also make unnecessary (most of the times) to
call self.client.<method> manually with an expectedStatus argument.

Finally, a special case is considered for the archive and qarchive
calls: if the file path they are given is relative, we consider it
relative to the test/ package and is interpreted as a resource name.
Even though the way this is implemented could be seen as "hacky", it
brings an huge benefit in terms of readability.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>