Removing unnecessary tmp config files
Unit test writers shouldn't need to write their own temporary unit tests
files; instead they only need to pass down the base configuration file
to use (if different from the default) and a set of configuration
properties that need to be overridden, if any.

This was already possible for prepExtSrv, but not in prepCluster, so we
had to add the functionality there.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>