New _to_abs and resource functions
The _to_abs function is the basis for properly addressing test resources
that until now have been addressed as relative paths. Examples of these
are all the files under the src/ and ref/ directories of the test/
package. This relative addressing is the main motive of why tests need
to be run within the test/ directory. With this commit we are only
introducing this new function; future commits will start applying it.

ngamsTestSuite.resource is a convenience function that subclasses can
use without having to import new names from the ngamsTestLib module.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>