Basic test functions resource-friendly
These include prepExtSrv, db_aware_cfg, assert_ref_file and
checkQueryPlan (and prepDiskCfg, but this is currently unused).
All these functions receive or use internally a filename that refers
most of the time to a test resource; it's therefore easier to maintain
that ease for callers, and handle things internally.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>