this_dir not needed anymore
This was a workaround to be able to reference resources under test/ and
reference thet test/ directory itself. Now that we're properly fixed the
resource addressing situation we can remove this hack.

The only place we used this and that we need to still pay attention was
setting up the python path for the spawned servers. In that case we
really want the servers to see the test/ directory as a top-level
package (otherwise loading plug-ins that are under test/support wouldn't
work), and therefore we still perform that bit of logic.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>