Unit tests can be run outside test/ dir
With this commit we are chaning Travis to *not* move into the test/
directory before running the unit tests (and therefore letting pytest
find them all automatically), and adding the proper documentation bits
to let users know that this improvement has become available. We are
also moving down the .coveragerc file and adjusting it to its new home.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>