Correct disk ID calculation
This was a remnant from the not-so-old days when we still relied on
hardcoded paths for temporary files. We hadn't realised this one was
left because the status objects coming back from the commands were not
being checked. Now that we're switching to assert-by-default calls we
found out this problem and corrected it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>