Correctly resource export in ngamsServer package
When we moved the different NGAS commands to the ngamsServer.commands
package we forgot to correctly amend the installation of the associated
status_doc.txt resource file, which was also moved to this sub-package.
As a result actual NGAS installations (and not "development"
installations) were lacking this file, and failed to reply to a simple
STATUS command.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>