Using sendfile to POST files
If we recognize `data` as being a file object that points to an actual
file then we use sendfile (via the pysendfile module) to send it through
the socket that underlies the HTTP connection object. We still use
conn.request() to place the request headers and flush it, leaving only
the body up to sendfile.

This change should benefit the ngamsPClient class and command tool in
cases when files are being archived. Additionally the subscription
thread should also see benefits, as it also POSTs files to subscribers.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>