NGAS-150 column name matches data in QUERY cmd
For a long time simple 'SELECT * FROM ...' sql commands have been used
to satisfy some of the queries supported by the QUERY command. This of
course means that the order in which columns appear on each row of data
is not predictable, and might not match the column name we associate to
each position.

Historically the QUERY command didn't return the actual names of the
columns being queried, as used "col<N>" as placeholders. This bug is
therefore specific to this new capability, which appeared only in NGAS
11.0.

A unit test has been added that performs a simple check on this aspect.
It does so by assuming that if it finds certain values on a few columns
then column names and values are correctly aligned.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>