Correcting HELP command test
We cannot use the built-in assertions for the HELP command tests because
those are ultimately based on the ngamsPClient class, which always
interprets redirections as a mechanism through which an NGAS server
offloads a command into *another* NGAS server. The new redirection we
have in the HELP command is different though: it is intended to
redirect a *browser* to visit a *webpage*, something that the
ngamsPClient class is not prepared for.

To avoid these problem the test issues a raw HTTP request and checks
that what comes back is what the browser would expect. The test module
has also been simplified to avoid code duplication and remove
unnecessary comments.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>